refactoring de la mort
