+++ include/engine_common.hpp
@@ -4,72 +4,12 @@
 #include <cmath>
 #include <cassert>
 #include <cstdio>
-#include <optional>
-#include <tuple>
 
 #include "common.hpp"
-#include "AbstractCollisionProblem.hpp"
-#include "counters.hpp"
+#include "problem.hpp"
 #include "dict.hpp"
 
 namespace mitm {
-
-class Parameters {
-public:
-    u64 nbytes_memory = 0;        /* how much RAM to use on each machine */
-    double difficulty = -1;       /* -log2(proportion of distinguished points). -1 == auto-choose */
-    double beta = 10;             /* use function variant for beta*w distinguished points */
-    int n_nodes = 1;              /* #hosts (with shared RAM) */
-    bool verbose = 1;             /* print progress information */
-
-    u64 threshold;                /* any integer less than this is a DP */
-    u64 dp_max_it;                /* how many iterations to find a DP */
-    u64 points_per_version;       /* #DP per version of the function */
-    u64 nslots;                   /* entries in the dict */
-
-    static double optimal_theta(double log2_w, int n)
-    {
-        return std::log2(2.25) + 0.5 * (log2_w - n);
-    }
-
-    void finalize(int n, u64 _nslots)
-    {
-        if (nbytes_memory == 0)
-            errx(1, "the size of the dictionnary must be specified");
-
-        nslots = _nslots;
-        double log2_w = std::log2(nslots * n_nodes);
-        double theta = optimal_theta(log2_w, n);
-        /* auto-choose the difficulty if not set */
-        if (difficulty < 0) {
-            difficulty = -theta;
-            if (difficulty < 0)
-                difficulty = 0;
-            if (verbose)
-                printf("AUTO-TUNING: setting difficulty %.2f\n", difficulty);
-        } else {
-            if (verbose)
-                printf("NOTICE: using difficulty=%.2f vs ``optimal''=%.2f\n", difficulty, theta);
-        }
-        threshold = std::pow(2., n - difficulty);
-        dp_max_it = 20 * std::pow(2., difficulty);
-        points_per_version = beta * nslots;
-
-        /* display warnings if problematic choices were made */
-        if (verbose && difficulty <= 0) {
-            printf("***** WARNING *****\n***** WARNING *****\n***** WARNING *****\n");
-            printf("---> zero difficulty (use the naive technique!)\n");
-            printf("***** WARNING *****\n***** WARNING *****\n***** WARNING *****\n");            
-        }
-        if (verbose && n - difficulty < log2_w) {
-            printf("***** WARNING *****\n***** WARNING *****\n***** WARNING *****\n");
-            printf("---> Too much memory (2^%.2f slots) but only 2^%.2f possible distinguished points\n",
-                log2_w, n - difficulty);
-            printf("***** WARNING *****\n***** WARNING *****\n***** WARNING *****\n");
-        }
-    }
-};
-
 
 /******************************************************************************/
 
@@ -110,7 +50,7 @@ static double sequential_benchmark(const ConcreteProblem& Pb)
  * Then return `true`. If the iterations limit is passed, returns `false`.
  */
 template<typename ConcreteProblem>
-std::optional<std::pair<u64,u64>> generate_dist_point(const ConcreteProblem& Pb, u64 i, const Parameters &params, u64 x)
+optional<pair<u64,u64>> generate_dist_point(const ConcreteProblem& Pb, u64 i, const Parameters &params, u64 x)
 {
     /* The probability, p, of NOT finding a distinguished point after the loop is
      * Let: theta := 2^-d
@@ -120,10 +60,10 @@ std::optional<std::pair<u64,u64>> generate_dist_point(const ConcreteProblem& Pb,
     for (u64 j = 0; j < params.dp_max_it; j++) {
         u64 y = Pb.mixf(i, x);
         if (is_distinguished_point(y, params.threshold))
-            return std::make_optional(std::pair(y, j+1));
+            return optional(pair(y, j+1));
         x = y;
     }
-    return std::nullopt; /* no distinguished point was found after too many iterations */
+    return nullopt; /* no distinguished point was found after too many iterations */
 }
 
 
@@ -133,7 +73,7 @@ std::optional<std::pair<u64,u64>> generate_dist_point(const ConcreteProblem& Pb,
  * add a drawing to illustrate this.
  */
 template<typename ConcreteProblem>
-std::optional<std::tuple<u64,u64,u64>> walk(const ConcreteProblem& Pb, u64 i, u64 x0, u64 len0, u64 x1, u64 len1)
+optional<tuple<u64,u64,u64>> walk(const ConcreteProblem& Pb, u64 i, u64 x0, u64 len0, u64 x1, u64 len1)
 {
   /****************************************************************************+
    *            walk the longest sequence until they are equal                 |
@@ -165,45 +105,45 @@ std::optional<std::tuple<u64,u64,u64>> walk(const ConcreteProblem& Pb, u64 i, u6
     /* First, do the outputs collide? If yes, return true and exit. */
     if (y0 == y1) {
       /* careful: x0 & x1 contain inputs before mixing */
-      return std::make_optional(std::tuple(x0, x1, y0));
+      return std::make_optional(tuple(x0, x1, y0));
     }
     x0 = y0;
     x1 = y1;
   }
-  return std::nullopt; /* we did not find a common point */
+  return nullopt; /* we did not find a common point */
 }
 
 
 template<class ConcreteProblem, class Counters>
-std::optional<std::tuple<u64,u64,u64>> process_distinguished_point(
-    const ConcreteProblem &Pb, Counters &ctr, Dict<std::pair<u64, u64>> &dict, 
+optional<tuple<u64,u64,u64>> process_distinguished_point(
+    const ConcreteProblem &Pb, Counters &ctr, PcsDict &dict, 
     u64 i, u64 start0, u64 end, u64 len0)
 {
-    auto probe = dict.pop_insert(end, std::pair(start0, len0));
+    auto probe = dict.pop_insert(end, start0, len0);
     if (not probe) {
         ctr.probe_failure();
-        return std::nullopt;
+        return nullopt;
     }
 
     auto [start1, len1] = *probe;
     auto collision = walk(Pb, i, start0, len0, start1, len1);  
     if (not collision) {
         ctr.walk_failure();
-        return std::nullopt;         /* robin-hood, or dict false positive */
+        return nullopt;         /* robin-hood, or dict false positive */
     }
 
     auto [x0, x1, y] = *collision;
     if (x0 == x1) {
         ctr.collision_failure();
-        return std::nullopt;    /* duh */
+        return nullopt;    /* duh */
     }
 
     ctr.found_collision();
     if (Pb.mix_good_pair(i, x0, x1)) {
         printf("\nFound golden collision!\n");
-        return std::make_optional(std::tuple(i, x0, x1));
+        return optional(tuple(i, x0, x1));
     }
-    return std::nullopt;
+    return nullopt;
 }
