+++ include/mpi/pcs_controller.hpp
@@ -5,17 +5,17 @@
 
 #include "common.hpp"
 #include "engine_common.hpp"
-#include "mpi_common.hpp"
+#include "mpi/common.hpp"
 
 namespace mitm {
 
 /* there is ONE controller process (of global rank 0) */
 
 template<typename ConcreteProblem>
-std::tuple<u64,u64,u64> controller(const ConcreteProblem& Pb, const MpiParameters &params, PRNG &prng)
+tuple<u64,u64,u64> controller(const ConcreteProblem& Pb, const MpiParameters &params, PRNG &prng)
 {
 	MpiCounters &ctr = Pb.ctr;
-    std::optional<std::tuple<u64,u64,u64>> solution;    /* (i, x0, x1)  */
+    optional<tuple<u64,u64,u64>> solution;    /* (i, x0, x1)  */
 	u64 stop = 0;
 	ctr.ready(Pb.n, params.nslots);
 	
@@ -63,7 +63,7 @@ std::tuple<u64,u64,u64> controller(const ConcreteProblem& Pb, const MpiParameter
 
 				case TAG_SOLUTION:
 					assert(i == buffer[0]);
-					solution = std::make_optional(std::tuple(buffer[0], buffer[1], buffer[2]));
+					solution = optional(tuple(buffer[0], buffer[1], buffer[2]));
 					stop = 1;
 			}
 		}
